+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Updated dependencies
 - Fixed issue in `pilet declaration` re-emitting remote types
+- Fixed issue in `piral-translate` where calling `addTranslation` multiple times caused previously added local translations to be removed
 
 ## 1.10.3 (March 24, 2026)